pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ascii-art"
version = "0.1.0"
description = "Generate ASCII art from images — plain, colored terminal, or HTML output."
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
dependencies = ["Pillow>=10.0"]

[project.scripts]
ascii-art = "ascii_art.cli:main"

[project.optional-dependencies]
dev = ["pytest>=7.0", "ruff>=0.4"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
